requirements.txt
gradio>=5.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
python-docx>=1.1.0
numpy>=1.24.0
google-cloud-firestore>=2.19.0
google-cloud-storage>=2.14.0
